requirements.txt
aiohttp==3.5.4
argcomplete==1.10.0
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorclass==2.2.0
idna==2.8
idna-ssl==1.1.0
macaroonbakery==1.2.3
multidict==4.5.2
oauthlib==3.0.1
protobuf==3.8.0
pycparser==2.19
pymacaroons==0.13.0
pymongo==3.8.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-libmaas==0.6.3
pytz==2019.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
terminaltables==3.1.0
typing-extensions==3.7.2
urllib3==1.25.3
yarl==1.3.0
